requirements.txt
psutil==5.9.6
requests==2.31.0
schedule==1.2.0
cryptography==41.0.7
python-dotenv==1.0.0
flask==3.0.0
sqlalchemy==2.0.23
